Fix GetAllNamedTypes
